Adjusted formatting Added feature to also works with rsync

+++ storage/innobase/xtrabackup/test/t/PXB-1548_rsync_basic.sh
@@ -4,26 +4,36 @@
 
 . inc/common.sh
 
-require_server_version_higher_than 5.7.02
+require_server_version_higher_than 5.6.0
 
 MYSQLD_EXTRA_MY_CNF_OPTS="
 innodb_buffer_pool_filename=ib_buffer_pool
 "
+
 if [[ -f  $mysql_datadir/ib_buffer_pool ]]; then
     rm -f $mysql_datadir/ib_buffer_pool
 fi
+
 start_server
 
 load_sakila
 
 mkdir $topdir/backup
 
-xtrabackup --backup --dump-innodb-buffer-pool --dump-innodb-buffer-pool-pct=100 --target-dir=$topdir/backup
+xtrabackup --backup --dump-innodb-buffer-pool --rsync --target-dir=$topdir/backup
+
+
+if [[ ! -f  $mysql_datadir/ib_buffer_pool ]]; then
+    die "dump file doesnt exist"
+fi
 
 $MYSQL $MYSQL_ARGS -e \
-       "SHOW ENGINE INNODB STATUS\G" | grep "Database pages" \
-       | head -1 | awk '{print $3}' > $topdir/status1
-       
+       "SHOW GLOBAL STATUS LIKE 'Innodb_buffer_pool_dump_status';" > $topdir/message
+
+grep -q "dump completed at" $topdir/message \
+|| die "Could not find \"dump completed at\" message"
+
+cat $mysql_datadir/ib_buffer_pool | wc -l > $topdir/status1
 cat $topdir/backup/ib_buffer_pool | wc -l > $topdir/status2
 
 diff -q $topdir/status1 $topdir/status2 \
